# CI speedup: move CScope db build to the cache-kernels job