When running regression tests on freshly downloaded kernels, it is usually very slow b/c it needs to build the CScope database for each used kernel version.
    
In CI, this is done for every job in the matrix, despite the fact that building CScope database does not depend on LLVM version.
    
To speed things up, this PR moves building of the CScope db to the 'cache-kernels' job which downloads and caches the kernels. This is done by running `make cscope` for each downloaded kernel.
    
Adding some comparison of CI runtimes below.
    
```
Current master:
- with cached kernels:    30m 56s
- without cached kernels: 43m 43s (11m 18s in cache-kernels)
    
This PR:
- with cached kernels:    19m 2s
- without cached kernels: 40m 8s (21m 4s in cache-kernels)
```
    
When considering some measurement errors, we can see that it takes roughly the same time when CI needs to refetch the kernels but there's a significant speedup when using pre-cached kernels.

This also required a small change in one unit test.